Move all user-specific controllers under users/
