+++ src/controllers/users/coffee/uff.listener.ts
@@ -3,16 +3,16 @@ import { Events } from "discord.js";
 import {
   CooldownManager,
   useCooldown,
-} from "../../middleware/cooldown.middleware";
+} from "../../../middleware/cooldown.middleware";
 import {
   contentMatching,
   messageFrom,
-} from "../../middleware/filters.middleware";
+} from "../../../middleware/filters.middleware";
 import {
   ListenerSpec,
   MessageListenerBuilder,
-} from "../../types/listener.types";
-import { replySilently } from "../../utils/interaction.utils";
+} from "../../../types/listener.types";
+import { replySilently } from "../../../utils/interaction.utils";
 
 const cooldown = new CooldownManager({ type: "global", seconds: 600 });
