+++ src/controllers/users/cxtie/chat-revive.listener.ts
@@ -1,13 +1,13 @@
 import { Events, Message } from "discord.js";
 
-import getLogger from "../../logger";
+import getLogger from "../../../logger";
 import {
   CooldownManager,
   useCooldown,
-} from "../../middleware/cooldown.middleware";
-import { ListenerSpec, MessageListenerBuilder } from "../../types/listener.types";
-import { replySilently } from "../../utils/interaction.utils";
-import uids from "../../utils/uids.utils";
+} from "../../../middleware/cooldown.middleware";
+import { ListenerSpec, MessageListenerBuilder } from "../../../types/listener.types";
+import { replySilently } from "../../../utils/interaction.utils";
+import uids from "../../../utils/uids.utils";
 
 const log = getLogger(__filename);
