+++ src/controllers/users/cxtie/sniffs.listener.ts
@@ -1,13 +1,13 @@
 
-import getLogger from "../../logger";
+import getLogger from "../../../logger";
 import {
   CooldownManager,
   useCooldown,
-} from "../../middleware/cooldown.middleware";
-import { messageFrom } from "../../middleware/filters.middleware";
-import { MessageListenerBuilder } from "../../types/listener.types";
-import { replySilently } from "../../utils/interaction.utils";
-import { formatContext } from "../../utils/logging.utils";
+} from "../../../middleware/cooldown.middleware";
+import { messageFrom } from "../../../middleware/filters.middleware";
+import { MessageListenerBuilder } from "../../../types/listener.types";
+import { replySilently } from "../../../utils/interaction.utils";
+import { formatContext } from "../../../utils/logging.utils";
 
 const log = getLogger(__filename);
